+++ doc/changelog.rst
@@ -6,6 +6,7 @@ Development Version
 
 - DOC improve documentation website, add estimator flowchart.
 - TST improve test robustness.
+- ENH add batching over targets in ``predict_weighted_kernel_ridge``.
 
 Version 0.3.5
 ~~~~~~~~~~~~~